update mirate file

+++ db/migrate/20181120070000_create_sells.rb
@@ -4,9 +4,9 @@ class CreateSells < ActiveRecord::Migration[5.2]
       t.belongs_to :product ,index: true
       t.integer :product_num
       t.string :sale_path
-      t.double :price
+      t.decimal :price
       t.date :plan_time
-      t.double :sale_expense
+      t.decimal :sale_expense
       t.timestamps
     end
   end
